refactor(config): split provider config support seams
