+++ crates/app/src/config/provider.rs
@@ -3,6 +3,13 @@ use std::{collections::BTreeMap, env, path::PathBuf};
 use loong_contracts::SecretRef;
 use serde::{Deserialize, Deserializer, Serialize};
 
+mod catalog;
+
+use self::catalog::{
+    PROVIDER_PROFILES, build_provider_descriptor_feature, find_cross_routed_validation_profile,
+    provider_descriptor_aliases, provider_descriptor_env_aliases, provider_descriptor_headers,
+    provider_descriptor_region_variants,
+};
 use super::shared::{
     ConfigValidationIssue, EnvPointerValidationHint, default_loong_home, expand_path,
     validate_env_pointer_field, validate_secret_ref_env_pointer_field,
@@ -10,6 +17,7 @@ use super::shared::{
 use crate::secrets::{
     SecretLookup, has_configured_secret_ref, resolve_secret_lookup, secret_ref_env_name,
 };
+pub use self::catalog::parse_provider_kind_id;
 
 pub(crate) const GITHUB_COPILOT_EDITOR_VERSION: &str = "vscode/1.85.1";
 pub(crate) const GITHUB_COPILOT_EDITOR_PLUGIN_VERSION: &str = "copilot/1.155.0";
@@ -2582,1540 +2590,6 @@ impl ProviderEffectiveUrlValues {
     }
 }
 
-impl ProviderUrlValidationProfile {
-    fn matching_canonical_fingerprint(
-        self,
-        value: &str,
-        allow_host_only_base_match: bool,
-    ) -> Option<&'static str> {
-        let profile = self.kind.profile();
-        let base_fingerprint = profile.base_url;
-        let base_match = matches_base_url_validation_fingerprint(
-            value,
-            base_fingerprint,
-            allow_host_only_base_match,
-        );
-        if base_match {
-            return Some(base_fingerprint);
-        }
-
-        for fingerprint in self.extra_canonical_url_fingerprints {
-            let fingerprint_match = matches_url_validation_fingerprint(value, fingerprint);
-            if fingerprint_match {
-                return Some(fingerprint);
-            }
-        }
-
-        None
-    }
-
-    fn matches_required_path_fragment(self, value: &str) -> bool {
-        for fragment in self.required_path_fragments {
-            let fragment_match = url_contains_validation_fragment(value, fragment);
-            if fragment_match {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    fn matches_forbidden_path_fragment(self, value: &str) -> bool {
-        for fragment in self.forbidden_path_exceptions {
-            let exception_match = url_contains_validation_fragment(value, fragment);
-            if exception_match {
-                return false;
-            }
-        }
-
-        for fragment in self.forbidden_path_fragments {
-            let fragment_match = url_contains_validation_fragment(value, fragment);
-            if fragment_match {
-                return true;
-            }
-        }
-
-        false
-    }
-}
-
-impl ProviderKind {
-    pub fn all_sorted() -> &'static [ProviderKind] {
-        &PROVIDER_KIND_ORDER
-    }
-
-    fn url_validation_profile(self) -> Option<&'static ProviderUrlValidationProfile> {
-        PROVIDER_URL_VALIDATION_PROFILES
-            .iter()
-            .find(|profile| profile.kind == self)
-    }
-
-    pub fn as_str(self) -> &'static str {
-        self.profile().id
-    }
-
-    pub fn display_name(self) -> &'static str {
-        match self {
-            ProviderKind::Anthropic => "Anthropic",
-            ProviderKind::Bedrock => "Bedrock",
-            ProviderKind::Byteplus => "BytePlus",
-            ProviderKind::ByteplusCoding => "BytePlus Coding",
-            ProviderKind::Cerebras => "Cerebras",
-            ProviderKind::CloudflareAiGateway => "Cloudflare AI Gateway",
-            ProviderKind::Cohere => "Cohere",
-            ProviderKind::Custom => "Custom",
-            ProviderKind::Deepseek => "DeepSeek",
-            ProviderKind::Fireworks => "Fireworks",
-            ProviderKind::Gemini => "Gemini",
-            ProviderKind::GithubCopilot => "GitHub Copilot",
-            ProviderKind::Groq => "Groq",
-            ProviderKind::Kimi => "Kimi",
-            ProviderKind::KimiCoding => "Kimi Coding",
-            ProviderKind::Mistral => "Mistral",
-            ProviderKind::Minimax => "MiniMax",
-            ProviderKind::Novita => "Novita",
-            ProviderKind::Nvidia => "NVIDIA",
-            ProviderKind::Llamacpp => "llama.cpp",
-            ProviderKind::LmStudio => "LM Studio",
-            ProviderKind::Ollama => "Ollama",
-            ProviderKind::Openai => "OpenAI",
-            ProviderKind::OpencodeZen => "OpenCode Zen",
-            ProviderKind::OpencodeGo => "OpenCode Go",
-            ProviderKind::Openrouter => "OpenRouter",
-            ProviderKind::Perplexity => "Perplexity",
-            ProviderKind::Qianfan => "Qianfan",
-            ProviderKind::Qwen => "Qwen",
-            ProviderKind::BailianCoding => "Bailian Coding",
-            ProviderKind::Sambanova => "SambaNova",
-            ProviderKind::Sglang => "SGLang",
-            ProviderKind::Siliconflow => "SiliconFlow",
-            ProviderKind::Stepfun => "StepFun",
-            ProviderKind::StepPlan => "Step Plan",
-            ProviderKind::Together => "Together",
-            ProviderKind::Venice => "Venice",
-            ProviderKind::VercelAiGateway => "Vercel AI Gateway",
-            ProviderKind::Vllm => "vLLM",
-            ProviderKind::Volcengine => "Volcengine",
-            ProviderKind::VolcengineCoding => "Volcengine Coding",
-            ProviderKind::Xai => "xAI",
-            ProviderKind::Xiaomi => "Xiaomi",
-            ProviderKind::Zai => "Z.ai",
-            ProviderKind::Zhipu => "Zhipu",
-        }
-    }
-
-    pub fn parse(raw: &str) -> Option<Self> {
-        parse_provider_kind_id(raw)
-    }
-
-    pub fn profile(self) -> &'static ProviderProfile {
-        let [
-            anthropic,
-            bailian_coding,
-            bedrock,
-            byteplus,
-            byteplus_coding,
-            cerebras,
-            cloudflare_ai_gateway,
-            cohere,
-            custom,
-            deepseek,
-            fireworks,
-            gemini,
-            github_copilot,
-            groq,
-            kimi,
-            kimi_coding,
-            llamacpp,
-            lm_studio,
-            mistral,
-            minimax,
-            novita,
-            nvidia,
-            ollama,
-            openai,
-            opencode_go,
-            opencode_zen,
-            openrouter,
-            perplexity,
-            qianfan,
-            qwen,
-            sambanova,
-            sglang,
-            siliconflow,
-            stepfun,
-            step_plan,
-            together,
-            venice,
-            vercel_ai_gateway,
-            vllm,
-            volcengine,
-            volcengine_coding,
-            xai,
-            xiaomi,
-            zai,
-            zhipu,
-        ] = &PROVIDER_PROFILES;
-
-        match self {
-            ProviderKind::Anthropic => anthropic,
-            ProviderKind::BailianCoding => bailian_coding,
-            ProviderKind::Bedrock => bedrock,
-            ProviderKind::Byteplus => byteplus,
-            ProviderKind::ByteplusCoding => byteplus_coding,
-            ProviderKind::Cerebras => cerebras,
-            ProviderKind::CloudflareAiGateway => cloudflare_ai_gateway,
-            ProviderKind::Cohere => cohere,
-            ProviderKind::Custom => custom,
-            ProviderKind::Deepseek => deepseek,
-            ProviderKind::Fireworks => fireworks,
-            ProviderKind::Gemini => gemini,
-            ProviderKind::GithubCopilot => github_copilot,
-            ProviderKind::Groq => groq,
-            ProviderKind::Kimi => kimi,
-            ProviderKind::KimiCoding => kimi_coding,
-            ProviderKind::Llamacpp => llamacpp,
-            ProviderKind::LmStudio => lm_studio,
-            ProviderKind::Mistral => mistral,
-            ProviderKind::Minimax => minimax,
-            ProviderKind::Novita => novita,
-            ProviderKind::Nvidia => nvidia,
-            ProviderKind::Ollama => ollama,
-            ProviderKind::Openai => openai,
-            ProviderKind::OpencodeZen => opencode_zen,
-            ProviderKind::OpencodeGo => opencode_go,
-            ProviderKind::Openrouter => openrouter,
-            ProviderKind::Perplexity => perplexity,
-            ProviderKind::Qianfan => qianfan,
-            ProviderKind::Qwen => qwen,
-            ProviderKind::Sambanova => sambanova,
-            ProviderKind::Sglang => sglang,
-            ProviderKind::Siliconflow => siliconflow,
-            ProviderKind::Stepfun => stepfun,
-            ProviderKind::StepPlan => step_plan,
-            ProviderKind::Together => together,
-            ProviderKind::Venice => venice,
-            ProviderKind::VercelAiGateway => vercel_ai_gateway,
-            ProviderKind::Vllm => vllm,
-            ProviderKind::Volcengine => volcengine,
-            ProviderKind::VolcengineCoding => volcengine_coding,
-            ProviderKind::Xai => xai,
-            ProviderKind::Xiaomi => xiaomi,
-            ProviderKind::Zai => zai,
-            ProviderKind::Zhipu => zhipu,
-        }
-    }
-
-    pub fn auth_scheme(self) -> ProviderAuthScheme {
-        self.profile().auth_scheme
-    }
-
-    pub fn protocol_family(self) -> ProviderProtocolFamily {
-        self.profile().protocol_family
-    }
-
-    pub fn feature_family(self) -> ProviderFeatureFamily {
-        self.profile().feature_family
-    }
-
-    pub fn default_headers(self) -> &'static [(&'static str, &'static str)] {
-        self.profile().default_headers
-    }
-
-    pub fn default_api_key_env(self) -> Option<&'static str> {
-        self.profile().default_api_key_env
-    }
-
-    pub fn api_key_env_aliases(self) -> &'static [&'static str] {
-        self.profile().api_key_env_aliases
-    }
-
-    pub fn default_user_agent(self) -> Option<&'static str> {
-        self.profile().default_user_agent
-    }
-
-    pub fn default_oauth_access_token_env(self) -> Option<&'static str> {
-        self.profile().default_oauth_access_token_env
-    }
-
-    pub fn oauth_access_token_env_aliases(self) -> &'static [&'static str] {
-        self.profile().oauth_access_token_env_aliases
-    }
-
-    pub fn auth_optional(self) -> bool {
-        matches!(
-            self,
-            ProviderKind::Llamacpp
-                | ProviderKind::LmStudio
-                | ProviderKind::Ollama
-                | ProviderKind::Sglang
-                | ProviderKind::Vllm
-        )
-    }
-
-    pub fn model_probe_auth_optional(self) -> bool {
-        self.auth_optional()
-            || matches!(self, ProviderKind::Cerebras | ProviderKind::VercelAiGateway)
-    }
-
-    pub fn allowed_reasoning_efforts(self) -> Option<&'static [ReasoningEffort]> {
-        if self == ProviderKind::Cohere {
-            Some(COHERE_REASONING_EFFORTS)
-        } else if self.feature_family() == ProviderFeatureFamily::Volcengine {
-            Some(ARK_REASONING_EFFORTS)
-        } else {
-            None
-        }
-    }
-
-    pub fn supports_reasoning_effort(self, effort: ReasoningEffort) -> bool {
-        self.allowed_reasoning_efforts()
-            .is_none_or(|allowed| allowed.contains(&effort))
-    }
-
-    pub fn prefers_max_completion_tokens(self) -> bool {
-        matches!(self, ProviderKind::Openai | ProviderKind::Cerebras)
-    }
-
-    pub fn preferred_token_limit_field_id(self) -> &'static str {
-        if self.prefers_max_completion_tokens() {
-            "max_completion_tokens"
-        } else {
-            "max_tokens"
-        }
-    }
-
-    pub fn requires_custom_base_url(self) -> bool {
-        matches!(
-            self,
-            ProviderKind::CloudflareAiGateway | ProviderKind::Custom
-        )
-    }
-
-    pub fn configuration_hint(self) -> Option<&'static str> {
-        if self == ProviderKind::Bedrock {
-            Some(
-                "set `BEDROCK_AWS_REGION`/`AWS_REGION`/`AWS_DEFAULT_REGION` or replace `<region>` in `provider.base_url` with your Bedrock runtime region",
-            )
-        } else if self == ProviderKind::CloudflareAiGateway {
-            Some(
-                "replace `<account_id>` and `<gateway_name>` in `provider.base_url` with your real Cloudflare AI Gateway path",
-            )
-        } else if self == ProviderKind::Custom {
-            Some(
-                "replace `<openai-compatible-host>` in `provider.base_url` with your real OpenAI-compatible endpoint root such as `https://api.example.com/v1`",
-            )
-        } else {
-            None
-        }
-    }
-
-    fn region_endpoint_guide(self) -> Option<ProviderRegionEndpointGuide> {
-        let profile = self.profile();
-        match self {
-            ProviderKind::Kimi => Some(ProviderRegionEndpointGuide {
-                family_label: "Moonshot Kimi",
-                default_variant: ProviderRegionEndpointVariant {
-                    label: "CN",
-                    base_url: profile.base_url,
-                },
-                alternate_variant: ProviderRegionEndpointVariant {
-                    label: "Global",
-                    base_url: "https://api.moonshot.ai",
-                },
-            }),
-            ProviderKind::Minimax => Some(ProviderRegionEndpointGuide {
-                family_label: "MiniMax",
-                default_variant: ProviderRegionEndpointVariant {
-                    label: "CN",
-                    base_url: profile.base_url,
-                },
-                alternate_variant: ProviderRegionEndpointVariant {
-                    label: "Global",
-                    base_url: "https://api.minimax.io",
-                },
-            }),
-            ProviderKind::Zai => Some(ProviderRegionEndpointGuide {
-                family_label: "Z.ai / BigModel",
-                default_variant: ProviderRegionEndpointVariant {
-                    label: "Global",
-                    base_url: profile.base_url,
-                },
-                alternate_variant: ProviderRegionEndpointVariant {
-                    label: "CN",
-                    base_url: "https://open.bigmodel.cn",
-                },
-            }),
-            ProviderKind::Zhipu => Some(ProviderRegionEndpointGuide {
-                family_label: "Z.ai / BigModel",
-                default_variant: ProviderRegionEndpointVariant {
-                    label: "CN",
-                    base_url: profile.base_url,
-                },
-                alternate_variant: ProviderRegionEndpointVariant {
-                    label: "Global",
-                    base_url: "https://api.z.ai",
-                },
-            }),
-            ProviderKind::Stepfun | ProviderKind::StepPlan => Some(ProviderRegionEndpointGuide {
-                family_label: "Stepfun",
-                default_variant: ProviderRegionEndpointVariant {
-                    label: "CN",
-                    base_url: profile.base_url,
-                },
-                alternate_variant: ProviderRegionEndpointVariant {
-                    label: "Global",
-                    base_url: "https://api.stepfun.ai",
-                },
-            }),
-            ProviderKind::Anthropic
-            | ProviderKind::Bedrock
-            | ProviderKind::Byteplus
-            | ProviderKind::ByteplusCoding
-            | ProviderKind::Cerebras
-            | ProviderKind::CloudflareAiGateway
-            | ProviderKind::Cohere
-            | ProviderKind::Custom
-            | ProviderKind::Deepseek
-            | ProviderKind::Fireworks
-            | ProviderKind::Gemini
-            | ProviderKind::Groq
-            | ProviderKind::GithubCopilot
-            | ProviderKind::KimiCoding
-            | ProviderKind::Llamacpp
-            | ProviderKind::LmStudio
-            | ProviderKind::Mistral
-            | ProviderKind::Novita
-            | ProviderKind::Nvidia
-            | ProviderKind::Ollama
-            | ProviderKind::Openai
-            | ProviderKind::OpencodeZen
-            | ProviderKind::OpencodeGo
-            | ProviderKind::Openrouter
-            | ProviderKind::Perplexity
-            | ProviderKind::Qianfan
-            | ProviderKind::Qwen
-            | ProviderKind::BailianCoding
-            | ProviderKind::Sambanova
-            | ProviderKind::Sglang
-            | ProviderKind::Siliconflow
-            | ProviderKind::Together
-            | ProviderKind::Venice
-            | ProviderKind::VercelAiGateway
-            | ProviderKind::Vllm
-            | ProviderKind::Volcengine
-            | ProviderKind::VolcengineCoding
-            | ProviderKind::Xai
-            | ProviderKind::Xiaomi => None,
-        }
-    }
-
-    pub const fn default_model(self) -> Option<&'static str> {
-        if matches!(self, ProviderKind::KimiCoding) {
-            Some("kimi-for-coding")
-        } else {
-            None
-        }
-    }
-
-    pub const fn recommended_onboarding_model(self) -> Option<&'static str> {
-        if matches!(self, ProviderKind::Deepseek) {
-            Some("deepseek-chat")
-        } else if matches!(self, ProviderKind::Minimax) {
-            Some("MiniMax-M2.7")
-        } else if matches!(self, ProviderKind::Xiaomi) {
-            Some("mimo-v2-pro")
-        } else {
-            None
-        }
-    }
-
-    pub fn region_endpoint_info(self) -> Option<ProviderRegionEndpointInfo> {
-        let guide = self.region_endpoint_guide()?;
-        let family_label = if matches!(self, ProviderKind::Zai | ProviderKind::Zhipu) {
-            "Z.ai"
-        } else {
-            guide.family_label
-        };
-        let variants = vec![
-            RegionVariant {
-                label: guide.default_variant.label,
-                base_url: guide.default_variant.base_url,
-            },
-            RegionVariant {
-                label: guide.alternate_variant.label,
-                base_url: guide.alternate_variant.base_url,
-            },
-        ];
-        Some(ProviderRegionEndpointInfo {
-            family_label,
-            variants,
-        })
-    }
-}
-
-const NON_CODING_ARK_FORBIDDEN_PATH_FRAGMENTS: [&str; 1] = ["/api/coding"];
-const CODING_ARK_REQUIRED_PATH_FRAGMENTS: [&str; 1] = ["/api/coding/v3"];
-const CODING_ARK_FORBIDDEN_PATH_FRAGMENTS: [&str; 2] = ["/api/v3", "/api/coding"];
-const CODING_ARK_FORBIDDEN_PATH_EXCEPTIONS: [&str; 1] = ["/api/coding/v3"];
-const VOLCENGINE_STANDARD_CANONICAL_URL_FINGERPRINTS: [&str; 1] =
-    ["https://ark.cn-beijing.volces.com/api/v3"];
-
-const PROVIDER_URL_VALIDATION_PROFILES: [ProviderUrlValidationProfile; 4] = [
-    ProviderUrlValidationProfile {
-        kind: ProviderKind::Byteplus,
-        extra_canonical_url_fingerprints: &[],
-        required_path_fragments: &[],
-        forbidden_path_fragments: &NON_CODING_ARK_FORBIDDEN_PATH_FRAGMENTS,
-        forbidden_path_exceptions: &[],
-        route_expectation: "the standard BytePlus ModelArk route under `/api/v3`",
-        path_validation_hint: "byteplus uses the standard ModelArk path and should not target `/api/coding` or `/api/coding/v3`; switch to `kind = \"byteplus_coding\"` for the dedicated OpenAI-compatible Coding Plan endpoint",
-    },
-    ProviderUrlValidationProfile {
-        kind: ProviderKind::ByteplusCoding,
-        extra_canonical_url_fingerprints: &[],
-        required_path_fragments: &CODING_ARK_REQUIRED_PATH_FRAGMENTS,
-        forbidden_path_fragments: &CODING_ARK_FORBIDDEN_PATH_FRAGMENTS,
-        forbidden_path_exceptions: &CODING_ARK_FORBIDDEN_PATH_EXCEPTIONS,
-        route_expectation: "the dedicated BytePlus Coding path under `/api/coding/v3`",
-        path_validation_hint: "byteplus_coding must use the dedicated BytePlus Coding path under `/api/coding/v3`; do not point it at the unsupported Anthropic-compatible `/api/coding` or generic `/api/v3` ModelArk endpoints because that bypasses Coding Plan quota and can incur standard model charges",
-    },
-    ProviderUrlValidationProfile {
-        kind: ProviderKind::Volcengine,
-        extra_canonical_url_fingerprints: &VOLCENGINE_STANDARD_CANONICAL_URL_FINGERPRINTS,
-        required_path_fragments: &[],
-        forbidden_path_fragments: &NON_CODING_ARK_FORBIDDEN_PATH_FRAGMENTS,
-        forbidden_path_exceptions: &[],
-        route_expectation: "the standard Volcengine Ark route under `/api/v3`",
-        path_validation_hint: "volcengine uses the standard Ark API path under `/api/v3` and should not target `/api/coding` or `/api/coding/v3`; switch to `kind = \"volcengine_coding\"` for the dedicated OpenAI-compatible Coding Plan endpoint",
-    },
-    ProviderUrlValidationProfile {
-        kind: ProviderKind::VolcengineCoding,
-        extra_canonical_url_fingerprints: &[],
-        required_path_fragments: &CODING_ARK_REQUIRED_PATH_FRAGMENTS,
-        forbidden_path_fragments: &CODING_ARK_FORBIDDEN_PATH_FRAGMENTS,
-        forbidden_path_exceptions: &CODING_ARK_FORBIDDEN_PATH_EXCEPTIONS,
-        route_expectation: "the dedicated Volcengine Coding Plan path under `/api/coding/v3`",
-        path_validation_hint: "volcengine_coding must use the dedicated Volcengine Coding Plan path under `/api/coding/v3`; do not point it at the Anthropic-compatible `/api/coding` or generic `/api/v3` Ark endpoints because that bypasses Coding Plan quota and can incur standard charges",
-    },
-];
-
-fn find_cross_routed_validation_profile(
-    current_kind: ProviderKind,
-    source_value: &str,
-    allow_host_only_base_match: bool,
-) -> Option<(&'static ProviderUrlValidationProfile, &'static str)> {
-    let mut best_match: Option<(&'static ProviderUrlValidationProfile, &'static str)> = None;
-
-    for profile in &PROVIDER_URL_VALIDATION_PROFILES {
-        let candidate_kind = profile.kind;
-        if candidate_kind == current_kind {
-            continue;
-        }
-
-        let matching_fingerprint =
-            profile.matching_canonical_fingerprint(source_value, allow_host_only_base_match);
-        if let Some(fingerprint) = matching_fingerprint {
-            let mut should_replace = true;
-            if let Some((_, best_fingerprint)) = best_match {
-                should_replace = fingerprint.len() > best_fingerprint.len();
-            }
-            if should_replace {
-                best_match = Some((profile, fingerprint));
-            }
-        }
-    }
-
-    best_match
-}
-
-fn matches_url_validation_fingerprint(value: &str, fingerprint: &str) -> bool {
-    let normalized_value = normalize_url_validation_value(value);
-    let normalized_fingerprint = normalize_url_validation_value(fingerprint);
-
-    matches_region_endpoint_url(normalized_value.as_str(), normalized_fingerprint.as_str())
-}
-
-fn matches_base_url_validation_fingerprint(
-    value: &str,
-    fingerprint: &str,
-    allow_host_only_base_match: bool,
-) -> bool {
-    let fingerprint_has_non_root_path = url_has_non_root_path(fingerprint);
-    if fingerprint_has_non_root_path {
-        return matches_url_validation_fingerprint(value, fingerprint);
-    }
-    if !allow_host_only_base_match {
-        return false;
-    }
-    is_same_base_url(value, fingerprint)
-}
-
-fn url_contains_validation_fragment(value: &str, fragment: &str) -> bool {
-    let normalized_value = normalize_url_validation_value(value);
-    let normalized_fragment = normalize_url_validation_value(fragment);
-
-    normalized_value.contains(normalized_fragment.as_str())
-}
-
-fn url_has_non_root_path(value: &str) -> bool {
-    let trimmed = value.trim();
-    let after_scheme = trimmed
-        .split_once("://")
-        .map(|(_, remainder)| remainder)
-        .unwrap_or(trimmed);
-    let Some((_, path_with_query)) = after_scheme.split_once('/') else {
-        return false;
-    };
-    let raw_path = format!("/{path_with_query}");
-    let path = raw_path
-        .split(['?', '#'])
-        .next()
-        .unwrap_or(raw_path.as_str());
-    let normalized_path = path.trim_end_matches('/');
-
-    !normalized_path.is_empty()
-}
-
-fn normalize_url_validation_value(value: &str) -> String {
-    let trimmed = value.trim();
-    let trimmed = trimmed.trim_end_matches('/');
-
-    trimmed.to_ascii_lowercase()
-}
-
-fn provider_descriptor_aliases(profile: &ProviderProfile) -> Vec<String> {
-    let mut aliases = Vec::new();
-
-    for alias in profile.aliases {
-        let alias = (*alias).to_owned();
-        aliases.push(alias);
-    }
-
-    aliases
-}
-
-fn provider_descriptor_headers(profile: &ProviderProfile) -> Vec<ProviderDescriptorHeader> {
-    let mut headers = Vec::new();
-
-    for (name, value) in profile.default_headers {
-        let header = ProviderDescriptorHeader {
-            name: (*name).to_owned(),
-            value: (*value).to_owned(),
-        };
-        headers.push(header);
-    }
-
-    headers
-}
-
-fn provider_descriptor_env_aliases(raw_aliases: &[&str]) -> Vec<String> {
-    let mut aliases = Vec::new();
-
-    for raw_alias in raw_aliases {
-        let alias = (*raw_alias).to_owned();
-        aliases.push(alias);
-    }
-
-    aliases
-}
-
-fn build_provider_descriptor_feature(
-    feature_support: &ProviderFeatureSupportFacts,
-) -> ProviderDescriptorFeature {
-    let family = feature_support.family.as_str().to_owned();
-    let gate_name = feature_support.gate_name.to_owned();
-    let enabled_in_build = feature_support.enabled_in_build;
-    let disabled_message = feature_support.disabled_message.clone();
-
-    ProviderDescriptorFeature {
-        family,
-        gate_name,
-        enabled_in_build,
-        disabled_message,
-    }
-}
-
-fn provider_descriptor_region_variants(
-    region_endpoint_info: Option<ProviderRegionEndpointInfo>,
-) -> Vec<ProviderDescriptorRegionVariant> {
-    let mut variants = Vec::new();
-
-    let Some(region_endpoint_info) = region_endpoint_info else {
-        return variants;
-    };
-
-    for variant in region_endpoint_info.variants {
-        let descriptor_variant = ProviderDescriptorRegionVariant {
-            label: variant.label.to_owned(),
-            base_url: variant.base_url.to_owned(),
-        };
-        variants.push(descriptor_variant);
-    }
-
-    variants
-}
-
-pub fn parse_provider_kind_id(raw: &str) -> Option<ProviderKind> {
-    let normalized = raw.trim().to_ascii_lowercase();
-    if normalized.is_empty() {
-        return None;
-    }
-
-    for profile in &PROVIDER_PROFILES {
-        if normalized == profile.id {
-            return Some(profile.kind);
-        }
-        if profile.aliases.iter().any(|alias| normalized == *alias) {
-            return Some(profile.kind);
-        }
-    }
-
-    None
-}
-
-const PROVIDER_KIND_ORDER: [ProviderKind; 45] = [
-    ProviderKind::Anthropic,
-    ProviderKind::BailianCoding,
-    ProviderKind::Bedrock,
-    ProviderKind::Byteplus,
-    ProviderKind::ByteplusCoding,
-    ProviderKind::Cerebras,
-    ProviderKind::CloudflareAiGateway,
-    ProviderKind::Cohere,
-    ProviderKind::Custom,
-    ProviderKind::Deepseek,
-    ProviderKind::Fireworks,
-    ProviderKind::Gemini,
-    ProviderKind::GithubCopilot,
-    ProviderKind::Groq,
-    ProviderKind::Kimi,
-    ProviderKind::KimiCoding,
-    ProviderKind::Llamacpp,
-    ProviderKind::LmStudio,
-    ProviderKind::Mistral,
-    ProviderKind::Minimax,
-    ProviderKind::Novita,
-    ProviderKind::Nvidia,
-    ProviderKind::Ollama,
-    ProviderKind::Openai,
-    ProviderKind::OpencodeGo,
-    ProviderKind::OpencodeZen,
-    ProviderKind::Openrouter,
-    ProviderKind::Perplexity,
-    ProviderKind::Qianfan,
-    ProviderKind::Qwen,
-    ProviderKind::Sambanova,
-    ProviderKind::Sglang,
-    ProviderKind::Siliconflow,
-    ProviderKind::Stepfun,
-    ProviderKind::StepPlan,
-    ProviderKind::Together,
-    ProviderKind::Venice,
-    ProviderKind::VercelAiGateway,
-    ProviderKind::Vllm,
-    ProviderKind::Volcengine,
-    ProviderKind::VolcengineCoding,
-    ProviderKind::Xai,
-    ProviderKind::Xiaomi,
-    ProviderKind::Zai,
-    ProviderKind::Zhipu,
-];
-
-const PROVIDER_PROFILES: [ProviderProfile; 45] = [
-    ProviderProfile {
-        kind: ProviderKind::Anthropic,
-        id: "anthropic",
-        aliases: &["anthropic_compatible"],
-        base_url: "https://api.anthropic.com",
-        chat_completions_path: "/v1/messages",
-        models_path: Some("/v1/models"),
-        protocol_family: ProviderProtocolFamily::AnthropicMessages,
-        auth_scheme: ProviderAuthScheme::XApiKey,
-        default_headers: &ANTHROPIC_DEFAULT_HEADERS,
-        default_api_key_env: Some("ANTHROPIC_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Anthropic,
-    },
-    ProviderProfile {
-        kind: ProviderKind::BailianCoding,
-        id: "bailian_coding",
-        aliases: &["bailian_coding_compatible"],
-        base_url: "https://coding.dashscope.aliyuncs.com/v1",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("BAILIAN_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: Some("openclaw"),
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Bedrock,
-        id: "bedrock",
-        aliases: &["aws-bedrock", "aws_bedrock"],
-        base_url: "https://bedrock-runtime.<region>.amazonaws.com",
-        chat_completions_path: "/model/{modelId}/converse",
-        models_path: Some("https://bedrock.<region>.amazonaws.com/foundation-models"),
-        protocol_family: ProviderProtocolFamily::BedrockConverse,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("AWS_BEARER_TOKEN_BEDROCK"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Bedrock,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Byteplus,
-        id: "byteplus",
-        aliases: &["byteplus_compatible"],
-        base_url: "https://ark.ap-southeast.bytepluses.com/api/v3",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("BYTEPLUS_API_KEY"),
-        api_key_env_aliases: &["ARK_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Volcengine,
-    },
-    ProviderProfile {
-        kind: ProviderKind::ByteplusCoding,
-        id: "byteplus_coding",
-        aliases: &["byteplus_coding_compatible"],
-        base_url: "https://ark.ap-southeast.bytepluses.com/api/coding/v3",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("BYTEPLUS_API_KEY"),
-        api_key_env_aliases: &["ARK_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Volcengine,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Cerebras,
-        id: "cerebras",
-        aliases: &["cerebras_compatible"],
-        base_url: "https://api.cerebras.ai",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/public/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("CEREBRAS_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::CloudflareAiGateway,
-        id: "cloudflare_ai_gateway",
-        aliases: &[
-            "cloudflare-ai-gateway",
-            "cloudflare_ai",
-            "cloudflare-ai",
-            "cloudflare",
-        ],
-        base_url: "https://gateway.ai.cloudflare.com/v1/<account_id>/<gateway_name>/openai/compat",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("CLOUDFLARE_API_KEY"),
-        api_key_env_aliases: &["CLOUDFLARE_AI_GATEWAY_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Cohere,
-        id: "cohere",
-        aliases: &["cohere_compatible"],
-        base_url: "https://api.cohere.ai/compatibility",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("https://api.cohere.com/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("COHERE_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Custom,
-        id: "custom",
-        aliases: &["openai_custom", "custom_openai"],
-        base_url: "https://<openai-compatible-host>/v1",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("CUSTOM_PROVIDER_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Deepseek,
-        id: "deepseek",
-        aliases: &["deepseek_compatible"],
-        base_url: "https://api.deepseek.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("DEEPSEEK_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Fireworks,
-        id: "fireworks",
-        aliases: &["fireworks_compatible", "fireworks-ai"],
-        base_url: "https://api.fireworks.ai",
-        chat_completions_path: "/inference/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("FIREWORKS_API_KEY"),
-        api_key_env_aliases: &["FIREWORKS_AI_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Gemini,
-        id: "gemini",
-        aliases: &[
-            "gemini_compatible",
-            "google",
-            "google_gemini",
-            "google-gemini",
-        ],
-        base_url: "https://generativelanguage.googleapis.com/v1beta/openai",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("GEMINI_API_KEY"),
-        api_key_env_aliases: &["GOOGLE_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::GithubCopilot,
-        id: "github-copilot",
-        aliases: &["github_copilot", "copilot"],
-        base_url: "https://api.githubcopilot.com",
-        chat_completions_path: "/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &GITHUB_COPILOT_DEFAULT_HEADERS,
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: Some(GITHUB_COPILOT_USER_AGENT),
-        default_oauth_access_token_env: Some(GITHUB_COPILOT_OAUTH_TOKEN_ENV),
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Groq,
-        id: "groq",
-        aliases: &["groq_compatible"],
-        base_url: "https://api.groq.com",
-        chat_completions_path: "/openai/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("GROQ_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Kimi,
-        id: "kimi",
-        aliases: &["kimi_compatible", "moonshot", "moonshot_compatible"],
-        base_url: "https://api.moonshot.cn",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("MOONSHOT_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::KimiCoding,
-        id: "kimi_coding",
-        aliases: &["kimi_coding_compatible"],
-        base_url: "https://api.kimi.com",
-        chat_completions_path: "/coding/v1/chat/completions",
-        models_path: Some("/coding/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("KIMI_CODING_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: Some("KimiCLI/Loong"),
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Llamacpp,
-        id: "llamacpp",
-        aliases: &["llama.cpp", "llama_cpp"],
-        base_url: "http://127.0.0.1:8080",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::LmStudio,
-        id: "lm_studio",
-        aliases: &["lmstudio", "lm-studio"],
-        base_url: "http://127.0.0.1:1234",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Mistral,
-        id: "mistral",
-        aliases: &["mistral_compatible"],
-        base_url: "https://api.mistral.ai",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("MISTRAL_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Minimax,
-        id: "minimax",
-        aliases: &["minimax_compatible"],
-        base_url: "https://api.minimaxi.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("MINIMAX_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Novita,
-        id: "novita",
-        aliases: &["novita_compatible"],
-        base_url: "https://api.novita.ai",
-        chat_completions_path: "/v3/openai/chat/completions",
-        models_path: Some("/v3/openai/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("NOVITA_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Nvidia,
-        id: "nvidia",
-        aliases: &[
-            "nvidia_compatible",
-            "nvidia_nim",
-            "nvidia-nim",
-            "build.nvidia.com",
-        ],
-        base_url: "https://integrate.api.nvidia.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("NVIDIA_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Ollama,
-        id: "ollama",
-        aliases: &["ollama_compatible"],
-        base_url: "http://127.0.0.1:11434",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Openai,
-        id: "openai",
-        aliases: &["openai_compatible"],
-        base_url: "https://api.openai.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("OPENAI_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: Some("OPENAI_CODEX_OAUTH_TOKEN"),
-        oauth_access_token_env_aliases: &["OPENAI_OAUTH_ACCESS_TOKEN"],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::OpencodeGo,
-        id: "opencode_go",
-        aliases: &["opencode-go"],
-        base_url: OPENCODE_GO_BASE_URL,
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some(OPENCODE_API_KEY_ENV),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::OpencodeZen,
-        id: "opencode_zen",
-        aliases: &["opencode", "opencode-zen"],
-        base_url: OPENCODE_ZEN_BASE_URL,
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some(OPENCODE_API_KEY_ENV),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Openrouter,
-        id: "openrouter",
-        aliases: &["openrouter_compatible"],
-        base_url: "https://openrouter.ai",
-        chat_completions_path: "/api/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("OPENROUTER_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Perplexity,
-        id: "perplexity",
-        aliases: &["perplexity_compatible"],
-        base_url: "https://api.perplexity.ai",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("PERPLEXITY_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Qianfan,
-        id: "qianfan",
-        aliases: &["qianfan_compatible", "baidu"],
-        base_url: "https://qianfan.baidubce.com",
-        chat_completions_path: "/v2/chat/completions",
-        models_path: Some("/v2/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("QIANFAN_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Qwen,
-        id: "qwen",
-        aliases: &["qwen_compatible", "dashscope"],
-        base_url: "https://dashscope.aliyuncs.com",
-        chat_completions_path: "/compatible-mode/v1/chat/completions",
-        models_path: Some("/compatible-mode/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("DASHSCOPE_API_KEY"),
-        api_key_env_aliases: &["QWEN_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Sambanova,
-        id: "sambanova",
-        aliases: &["sambanova_compatible", "samba_nova"],
-        base_url: "https://api.sambanova.ai",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("SAMBANOVA_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Sglang,
-        id: "sglang",
-        aliases: &["sglang_compatible"],
-        base_url: "http://127.0.0.1:30000",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Siliconflow,
-        id: "siliconflow",
-        aliases: &["siliconflow_compatible"],
-        base_url: "https://api.siliconflow.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("SILICONFLOW_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Stepfun,
-        id: "stepfun",
-        aliases: &["stepfun_compatible"],
-        base_url: "https://api.stepfun.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("STEP_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::StepPlan,
-        id: "step_plan",
-        aliases: &["stepfun_step_plan"],
-        base_url: "https://api.stepfun.com",
-        chat_completions_path: "/step_plan/v1/chat/completions",
-        models_path: Some("/step_plan/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("STEP_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Together,
-        id: "together",
-        aliases: &["together_compatible", "together_ai", "together-ai"],
-        base_url: "https://api.together.xyz",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("TOGETHER_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Venice,
-        id: "venice",
-        aliases: &["venice_compatible"],
-        base_url: "https://api.venice.ai",
-        chat_completions_path: "/api/v1/chat/completions",
-        models_path: Some("/api/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("VENICE_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::VercelAiGateway,
-        id: "vercel_ai_gateway",
-        aliases: &["vercel-ai-gateway", "vercel_ai", "vercel-ai", "vercel"],
-        base_url: "https://ai-gateway.vercel.sh/v1",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("AI_GATEWAY_API_KEY"),
-        api_key_env_aliases: &["VERCEL_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Vllm,
-        id: "vllm",
-        aliases: &["vllm_compatible"],
-        base_url: "http://127.0.0.1:8000",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: None,
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Volcengine,
-        id: "volcengine",
-        aliases: &[
-            "volcengine_custom",
-            "volcengine_compatible",
-            "doubao",
-            "ark",
-        ],
-        base_url: "https://ark.cn-beijing.volces.com",
-        chat_completions_path: "/api/v3/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("ARK_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Volcengine,
-    },
-    ProviderProfile {
-        kind: ProviderKind::VolcengineCoding,
-        id: "volcengine_coding",
-        aliases: &["volcengine_coding_compatible"],
-        base_url: "https://ark.cn-beijing.volces.com/api/coding/v3",
-        chat_completions_path: "/chat/completions",
-        models_path: Some("/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("ARK_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::Volcengine,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Xai,
-        id: "xai",
-        aliases: &["xai_compatible", "grok"],
-        base_url: "https://api.x.ai",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/v1/language-models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("XAI_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Xiaomi,
-        id: "xiaomi",
-        aliases: &[
-            "xiaomi_compatible",
-            "xiaomi_mimo",
-            "xiaomi-mimo",
-            "mimo",
-            "mimo_compatible",
-        ],
-        base_url: "https://api.xiaomimimo.com",
-        chat_completions_path: "/v1/chat/completions",
-        models_path: Some("/v1/models"),
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("XIAOMI_API_KEY"),
-        api_key_env_aliases: &["MIMO_API_KEY", "XIAOMIMIMO_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Zai,
-        id: "zai",
-        aliases: &["zai_compatible", "z.ai"],
-        base_url: "https://api.z.ai",
-        chat_completions_path: "/api/paas/v4/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("ZAI_API_KEY"),
-        api_key_env_aliases: &[],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-    ProviderProfile {
-        kind: ProviderKind::Zhipu,
-        id: "zhipu",
-        aliases: &["zhipu_compatible", "glm", "bigmodel"],
-        base_url: "https://open.bigmodel.cn",
-        chat_completions_path: "/api/paas/v4/chat/completions",
-        models_path: None,
-        protocol_family: ProviderProtocolFamily::OpenAiChatCompletions,
-        auth_scheme: ProviderAuthScheme::Bearer,
-        default_headers: &[],
-        default_api_key_env: Some("ZHIPUAI_API_KEY"),
-        api_key_env_aliases: &["ZHIPU_API_KEY"],
-        default_user_agent: None,
-        default_oauth_access_token_env: None,
-        oauth_access_token_env_aliases: &[],
-        feature_family: ProviderFeatureFamily::OpenAiCompatible,
-    },
-];
-
 fn default_provider_model() -> String {
     "auto".to_owned()
 }
@@ -4488,880 +2962,4 @@ fn derive_responses_path(chat_path: &str) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::{Value, json};
-    use std::collections::{BTreeMap, BTreeSet};
-
-    use crate::test_support::ScopedEnv;
-    use loong_contracts::SecretRef;
-
-    fn encode_provider_descriptor(descriptor: &ProviderDescriptorDocument) -> Value {
-        serde_json::to_value(descriptor).expect("serialize provider descriptor document")
-    }
-
-    #[test]
-    fn provider_profile_lookup_matches_kind() {
-        for kind in ProviderKind::all_sorted() {
-            assert_eq!(kind.profile().kind, *kind);
-        }
-    }
-
-    #[test]
-    fn provider_profile_aliases_are_unique_and_do_not_shadow_ids() {
-        let mut provider_ids = BTreeSet::new();
-        for kind in ProviderKind::all_sorted() {
-            let profile = kind.profile();
-            let provider_id = profile.id;
-            let inserted = provider_ids.insert(provider_id);
-            assert!(inserted, "duplicate provider id detected: {provider_id}");
-        }
-
-        let mut alias_owners = BTreeMap::new();
-        for kind in ProviderKind::all_sorted() {
-            let profile = kind.profile();
-            let provider_id = profile.id;
-            for alias in profile.aliases {
-                let normalized_alias = alias.trim();
-                assert!(
-                    !normalized_alias.is_empty(),
-                    "provider `{provider_id}` contains an empty alias"
-                );
-                assert_ne!(
-                    normalized_alias, provider_id,
-                    "provider `{provider_id}` repeats its canonical id as an alias"
-                );
-                assert!(
-                    !provider_ids.contains(normalized_alias),
-                    "provider alias `{normalized_alias}` collides with a canonical provider id"
-                );
-
-                let previous_owner = alias_owners.insert(normalized_alias.to_owned(), provider_id);
-                assert!(
-                    previous_owner.is_none(),
-                    "provider alias `{normalized_alias}` is shared by `{}` and `{provider_id}`",
-                    previous_owner.unwrap_or_default()
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn provider_profile_aliases_round_trip_through_provider_kind_deserialization() {
-        for kind in ProviderKind::all_sorted() {
-            let profile = kind.profile();
-            let canonical_id = format!("\"{}\"", profile.id);
-            let parsed_canonical = serde_json::from_str::<ProviderKind>(canonical_id.as_str())
-                .expect("canonical provider id should deserialize");
-            assert_eq!(
-                parsed_canonical, *kind,
-                "canonical provider id should deserialize to its matching provider kind"
-            );
-
-            for alias in profile.aliases {
-                let raw_alias = format!("\"{alias}\"");
-                let parsed_alias = serde_json::from_str::<ProviderKind>(raw_alias.as_str())
-                    .expect("provider alias should deserialize");
-                assert_eq!(
-                    parsed_alias, *kind,
-                    "provider alias `{alias}` should deserialize to the same provider kind as parse_provider_kind_id"
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn provider_kind_serializes_github_copilot_using_canonical_hyphenated_id() {
-        let raw = serde_json::to_string(&ProviderKind::GithubCopilot)
-            .expect("provider kind should serialize");
-
-        assert_eq!(raw, "\"github-copilot\"");
-    }
-
-    #[test]
-    fn provider_feature_family_gate_messages_are_stable() {
-        let anthropic_message = ProviderFeatureFamily::Anthropic.disabled_message();
-        let bedrock_message = ProviderFeatureFamily::Bedrock.disabled_message();
-        let volcengine_message = ProviderFeatureFamily::Volcengine.disabled_message();
-        let openai_message = ProviderFeatureFamily::OpenAiCompatible.disabled_message();
-
-        assert_eq!(
-            anthropic_message,
-            "anthropic provider family is disabled (enable feature `provider-anthropic`)"
-        );
-        assert_eq!(
-            bedrock_message,
-            "bedrock provider family is disabled (enable feature `provider-bedrock`)"
-        );
-        assert_eq!(
-            volcengine_message,
-            "volcengine provider family is disabled (enable feature `provider-volcengine`)"
-        );
-        assert_eq!(
-            openai_message,
-            "openai-compatible provider family is disabled (enable feature `provider-openai`)"
-        );
-    }
-
-    #[test]
-    fn provider_profile_static_auth_hints_are_stable() {
-        let byteplus_hint = ProviderKind::ByteplusCoding
-            .profile()
-            .auth_guidance_hint()
-            .expect("byteplus coding should expose auth guidance");
-        let volcengine_hint = ProviderKind::Volcengine
-            .profile()
-            .auth_guidance_hint()
-            .expect("volcengine should expose auth guidance");
-        let bedrock_hint = ProviderKind::Bedrock
-            .profile()
-            .alternative_auth_configuration_hint()
-            .expect("bedrock should expose a SigV4 fallback hint");
-        let custom_hint = ProviderKind::Custom
-            .profile()
-            .alternative_auth_configuration_hint()
-            .expect("custom provider should expose header guidance");
-
-        assert!(byteplus_hint.contains("BytePlus"));
-        assert!(byteplus_hint.contains("BYTEPLUS_API_KEY"));
-        assert!(byteplus_hint.contains("Authorization: Bearer <BYTEPLUS_API_KEY>"));
-
-        assert!(volcengine_hint.contains("Volcengine"));
-        assert!(volcengine_hint.contains("ARK_API_KEY"));
-        assert!(volcengine_hint.contains("AK/SK request signing is not used"));
-
-        assert!(bedrock_hint.contains("AWS_ACCESS_KEY_ID"));
-        assert!(bedrock_hint.contains("AWS_SECRET_ACCESS_KEY"));
-        assert!(bedrock_hint.contains("BEDROCK_AWS_REGION"));
-        assert!(bedrock_hint.contains("AWS_REGION"));
-
-        assert!(custom_hint.contains("Authorization"));
-        assert!(custom_hint.contains("X-API-Key"));
-        assert!(custom_hint.contains("provider.headers"));
-    }
-
-    #[test]
-    fn provider_configuration_hint_prefers_canonical_cross_routing_guidance() {
-        let cases = [
-            (
-                ProviderKind::Byteplus,
-                "https://ark.cn-beijing.volces.com/api/coding/v3",
-                "kind = \"volcengine_coding\"",
-            ),
-            (
-                ProviderKind::ByteplusCoding,
-                "https://ark.cn-beijing.volces.com",
-                "kind = \"volcengine\"",
-            ),
-        ];
-
-        for (kind, base_url, expected_kind_hint) in cases {
-            let provider = ProviderConfig {
-                kind,
-                base_url: base_url.to_owned(),
-                base_url_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider
-                .configuration_hint()
-                .expect("cross-routed provider configs should surface a hint");
-
-            assert!(
-                hint.contains(expected_kind_hint),
-                "expected cross-routing hint for {kind:?} with {base_url}: {hint}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_requires_coding_path_even_for_proxy_hosts() {
-        let cases = [ProviderKind::ByteplusCoding, ProviderKind::VolcengineCoding];
-
-        for kind in cases {
-            let provider = ProviderConfig {
-                kind,
-                base_url: "https://proxy.example.com/openai/v1".to_owned(),
-                base_url_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider
-                .configuration_hint()
-                .expect("coding providers should reject proxy urls that drop the coding path");
-
-            assert!(
-                hint.contains("/api/coding/v3"),
-                "expected coding-path guidance for {kind:?}: {hint}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_rejects_non_coding_profiles_on_coding_proxy_paths() {
-        let cases = [
-            (ProviderKind::Byteplus, "kind = \"byteplus_coding\""),
-            (ProviderKind::Volcengine, "kind = \"volcengine_coding\""),
-        ];
-
-        for (kind, expected_kind_hint) in cases {
-            let provider = ProviderConfig {
-                kind,
-                base_url: "https://proxy.example.com/api/coding/v3".to_owned(),
-                base_url_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider
-                .configuration_hint()
-                .expect("non-coding profiles should reject coding proxy paths");
-
-            assert!(
-                hint.contains(expected_kind_hint),
-                "expected non-coding guidance for {kind:?}: {hint}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_allows_proxy_coding_routes_that_keep_the_required_path() {
-        let cases = [ProviderKind::ByteplusCoding, ProviderKind::VolcengineCoding];
-
-        for kind in cases {
-            let provider = ProviderConfig {
-                kind,
-                base_url: "https://proxy.example.com/api/coding/v3".to_owned(),
-                base_url_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider.configuration_hint();
-
-            assert!(
-                hint.is_none(),
-                "proxy coding routes should stay valid for {kind:?}: {hint:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_checks_explicit_endpoints_before_current_base_url() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::ByteplusCoding,
-            base_url: "https://ark.ap-southeast.bytepluses.com/api/coding/v3".to_owned(),
-            base_url_explicit: true,
-            endpoint: Some("https://ark.cn-beijing.volces.com/api/v3/chat/completions".to_owned()),
-            endpoint_explicit: true,
-            ..ProviderConfig::default()
-        };
-        let hint = provider
-            .configuration_hint()
-            .expect("explicit cross-routed endpoints should surface a hint");
-
-        assert!(hint.contains("kind = \"volcengine\""));
-    }
-
-    #[test]
-    fn provider_configuration_hint_requires_coding_path_for_explicit_endpoint_overrides() {
-        let cases = [ProviderKind::ByteplusCoding, ProviderKind::VolcengineCoding];
-
-        for kind in cases {
-            let provider = ProviderConfig {
-                kind,
-                endpoint: Some("https://proxy.example.com/openai/v1/chat/completions".to_owned()),
-                endpoint_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider
-                .configuration_hint()
-                .expect("explicit endpoint overrides should keep the coding path");
-
-            assert!(
-                hint.contains("/api/coding/v3"),
-                "expected coding-path guidance for {kind:?}: {hint}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_requires_coding_path_for_explicit_models_endpoints() {
-        let cases = [ProviderKind::ByteplusCoding, ProviderKind::VolcengineCoding];
-
-        for kind in cases {
-            let provider = ProviderConfig {
-                kind,
-                models_endpoint: Some("https://proxy.example.com/openai/v1/models".to_owned()),
-                models_endpoint_explicit: true,
-                ..ProviderConfig::default()
-            };
-            let hint = provider
-                .configuration_hint()
-                .expect("explicit models endpoints should keep the coding path");
-
-            assert!(
-                hint.contains("/api/coding/v3"),
-                "expected coding-path guidance for {kind:?}: {hint}"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_configuration_hint_ignores_non_canonical_host_only_proxy_paths() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Byteplus,
-            base_url: "https://ark.cn-beijing.volces.com/custom/proxy".to_owned(),
-            base_url_explicit: true,
-            ..ProviderConfig::default()
-        };
-        let hint = provider.configuration_hint();
-
-        assert!(
-            hint.is_none(),
-            "non-canonical host-only proxy paths should not claim a canonical cross-route: {hint:?}"
-        );
-    }
-
-    #[test]
-    fn provider_feature_support_facts_are_stable() {
-        let facts = ProviderFeatureFamily::Volcengine.support_facts();
-
-        assert_eq!(facts.family, ProviderFeatureFamily::Volcengine);
-        assert_eq!(facts.gate_name, "provider-volcengine");
-        assert_eq!(
-            facts.enabled_in_build,
-            ProviderFeatureFamily::Volcengine.is_enabled_in_build()
-        );
-        assert_eq!(
-            facts.disabled_message,
-            "volcengine provider family is disabled (enable feature `provider-volcengine`)"
-        );
-    }
-
-    #[test]
-    fn provider_support_facts_preserve_auth_guidance_and_missing_auth_message() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::ByteplusCoding,
-            ..ProviderConfig::default()
-        };
-
-        let support_facts = provider.support_facts();
-        let auth_support = support_facts.auth;
-        let guidance_hint = auth_support
-            .guidance_hint
-            .expect("byteplus coding should expose auth guidance");
-
-        assert!(auth_support.requires_explicit_configuration);
-        assert!(
-            auth_support
-                .hint_env_names
-                .contains(&"BYTEPLUS_API_KEY".to_owned())
-        );
-        assert!(guidance_hint.contains("BytePlus"));
-        assert!(guidance_hint.contains("BYTEPLUS_API_KEY"));
-        assert!(guidance_hint.contains("Authorization: Bearer <BYTEPLUS_API_KEY>"));
-        assert!(
-            auth_support
-                .missing_configuration_message
-                .contains("BYTEPLUS_API_KEY")
-        );
-        assert!(
-            auth_support
-                .missing_configuration_message
-                .contains("BytePlus")
-        );
-    }
-
-    #[test]
-    fn auth_hint_env_names_respect_explicit_api_key_env_binding_precedence() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Openai,
-            api_key: Some(SecretRef::Env {
-                env: "TEAM_OPENAI_KEY".to_owned(),
-            }),
-            ..ProviderConfig::default()
-        };
-
-        let env_names = provider.auth_hint_env_names();
-
-        assert_eq!(env_names, vec!["TEAM_OPENAI_KEY".to_owned()]);
-    }
-
-    #[test]
-    fn auth_hint_env_names_keep_api_key_fallback_after_explicit_oauth_env_binding() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Openai,
-            oauth_access_token: Some(SecretRef::Env {
-                env: "TEAM_OPENAI_OAUTH".to_owned(),
-            }),
-            ..ProviderConfig::default()
-        };
-
-        let env_names = provider.auth_hint_env_names();
-
-        assert_eq!(
-            env_names,
-            vec!["TEAM_OPENAI_OAUTH".to_owned(), "OPENAI_API_KEY".to_owned(),]
-        );
-    }
-
-    #[test]
-    fn provider_support_facts_preserve_region_endpoint_hints() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Minimax,
-            ..ProviderConfig::default()
-        };
-
-        let support_facts = provider.support_facts();
-        let region_endpoint_support = support_facts.region_endpoint;
-        let note = region_endpoint_support
-            .note
-            .expect("minimax should expose a region endpoint note");
-        let catalog_failure_hint = region_endpoint_support
-            .catalog_failure_hint
-            .expect("minimax should expose a catalog failure hint");
-        let request_failure_hint = region_endpoint_support
-            .request_failure_hint
-            .expect("minimax should expose a request failure hint");
-
-        assert!(note.contains("MiniMax region endpoint"));
-        assert!(note.contains("https://api.minimaxi.com"));
-        assert!(note.contains("https://api.minimax.io"));
-        assert!(catalog_failure_hint.contains("https://api.minimaxi.com"));
-        assert!(catalog_failure_hint.contains("https://api.minimax.io"));
-        assert!(request_failure_hint.contains("https://api.minimaxi.com"));
-        assert!(request_failure_hint.contains("https://api.minimax.io"));
-    }
-
-    #[test]
-    fn step_plan_provider_support_facts_expose_region_endpoint_variants() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::StepPlan,
-            ..ProviderConfig::default()
-        };
-
-        let support_facts = provider.support_facts();
-        let region_endpoint_support = support_facts.region_endpoint;
-        let note = region_endpoint_support
-            .note
-            .expect("step plan should expose a region endpoint note");
-        let catalog_failure_hint = region_endpoint_support
-            .catalog_failure_hint
-            .expect("step plan should expose a catalog failure hint");
-        let request_failure_hint = region_endpoint_support
-            .request_failure_hint
-            .expect("step plan should expose a request failure hint");
-
-        assert!(note.contains("Stepfun"));
-        assert!(note.contains("https://api.stepfun.com"));
-        assert!(note.contains("https://api.stepfun.ai"));
-        assert!(catalog_failure_hint.contains("https://api.stepfun.com"));
-        assert!(catalog_failure_hint.contains("https://api.stepfun.ai"));
-        assert!(request_failure_hint.contains("https://api.stepfun.com"));
-        assert!(request_failure_hint.contains("https://api.stepfun.ai"));
-    }
-
-    #[test]
-    fn provider_descriptor_document_preserves_x_api_key_contract_facts() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Anthropic,
-            ..ProviderConfig::default()
-        };
-
-        let descriptor = provider.descriptor_document();
-        let encoded = encode_provider_descriptor(&descriptor);
-        let hint_env_names = encoded["auth"]["hint_env_names"]
-            .as_array()
-            .expect("hint env names should be an array");
-
-        assert_eq!(
-            encoded["schema"]["version"],
-            json!(PROVIDER_DESCRIPTOR_SCHEMA_VERSION)
-        );
-        assert_eq!(encoded["schema"]["surface"], json!("provider_descriptor"));
-        assert_eq!(encoded["schema"]["purpose"], json!("internal_sdk_contract"));
-        assert_eq!(encoded["kind"], json!("anthropic"));
-        assert_eq!(encoded["display_name"], json!("Anthropic"));
-        assert_eq!(encoded["protocol_family"], json!("anthropic_messages"));
-        assert_eq!(encoded["feature"]["family"], json!("anthropic"));
-        assert_eq!(encoded["auth"]["scheme"], json!("x_api_key"));
-        assert_eq!(
-            encoded["auth"]["default_api_key_env"],
-            json!("ANTHROPIC_API_KEY")
-        );
-        assert_eq!(
-            encoded["auth"]["requires_explicit_configuration"],
-            json!(true)
-        );
-        assert!(
-            hint_env_names.contains(&json!("ANTHROPIC_API_KEY")),
-            "anthropic descriptor should surface the canonical x-api-key env hint"
-        );
-        assert_eq!(
-            encoded["default_headers"][0]["name"],
-            json!("anthropic-version")
-        );
-    }
-
-    #[test]
-    fn provider_descriptor_document_marks_auth_optional_profiles_without_required_envs() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Ollama,
-            ..ProviderConfig::default()
-        };
-
-        let descriptor = provider.descriptor_document();
-        let encoded = encode_provider_descriptor(&descriptor);
-
-        assert_eq!(encoded["kind"], json!("ollama"));
-        assert_eq!(encoded["auth"]["scheme"], json!("bearer"));
-        assert_eq!(encoded["auth"]["auth_optional"], json!(true));
-        assert_eq!(encoded["auth"]["model_probe_auth_optional"], json!(true));
-        assert_eq!(
-            encoded["auth"]["requires_explicit_configuration"],
-            json!(false)
-        );
-        assert_eq!(encoded["auth"]["hint_env_names"], json!([]));
-        assert_eq!(encoded["region_endpoint"]["variants"], json!([]));
-    }
-
-    #[test]
-    fn provider_descriptor_document_prefers_dynamic_configuration_hint() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Custom,
-            ..ProviderConfig::default()
-        };
-
-        let descriptor = provider.descriptor_document();
-        let encoded = encode_provider_descriptor(&descriptor);
-        let configuration_hint = encoded["configuration_hint"]
-            .as_str()
-            .expect("custom descriptor should expose a configuration hint");
-
-        assert!(configuration_hint.contains("tenant-scoped base_url configuration"));
-        assert!(configuration_hint.contains("current template"));
-        assert!(configuration_hint.contains("https://<openai-compatible-host>/v1"));
-    }
-
-    #[test]
-    fn provider_descriptor_document_preserves_region_endpoint_variants_and_hints() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::Minimax,
-            ..ProviderConfig::default()
-        };
-
-        let descriptor = provider.descriptor_document();
-        let encoded = encode_provider_descriptor(&descriptor);
-        let note = encoded["region_endpoint"]["note"]
-            .as_str()
-            .expect("minimax descriptor should expose a region note");
-        let catalog_failure_hint = encoded["region_endpoint"]["catalog_failure_hint"]
-            .as_str()
-            .expect("minimax descriptor should expose a catalog failure hint");
-        let request_failure_hint = encoded["region_endpoint"]["request_failure_hint"]
-            .as_str()
-            .expect("minimax descriptor should expose a request failure hint");
-
-        assert_eq!(encoded["region_endpoint"]["family_label"], json!("MiniMax"));
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][0]["label"],
-            json!("CN")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][0]["base_url"],
-            json!("https://api.minimaxi.com")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][1]["label"],
-            json!("Global")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][1]["base_url"],
-            json!("https://api.minimax.io")
-        );
-        assert!(note.contains("MiniMax region endpoint"));
-        assert!(catalog_failure_hint.contains("https://api.minimaxi.com"));
-        assert!(catalog_failure_hint.contains("https://api.minimax.io"));
-        assert!(request_failure_hint.contains("https://api.minimaxi.com"));
-        assert!(request_failure_hint.contains("https://api.minimax.io"));
-    }
-
-    #[test]
-    fn step_plan_descriptor_document_preserves_region_endpoint_variants_and_hints() {
-        let provider = ProviderConfig {
-            kind: ProviderKind::StepPlan,
-            ..ProviderConfig::default()
-        };
-
-        let descriptor = provider.descriptor_document();
-        let encoded = encode_provider_descriptor(&descriptor);
-        let note = encoded["region_endpoint"]["note"]
-            .as_str()
-            .expect("step plan descriptor should expose a region note");
-        let catalog_failure_hint = encoded["region_endpoint"]["catalog_failure_hint"]
-            .as_str()
-            .expect("step plan descriptor should expose a catalog failure hint");
-        let request_failure_hint = encoded["region_endpoint"]["request_failure_hint"]
-            .as_str()
-            .expect("step plan descriptor should expose a request failure hint");
-
-        assert_eq!(encoded["region_endpoint"]["family_label"], json!("Stepfun"));
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][0]["label"],
-            json!("CN")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][0]["base_url"],
-            json!("https://api.stepfun.com")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][1]["label"],
-            json!("Global")
-        );
-        assert_eq!(
-            encoded["region_endpoint"]["variants"][1]["base_url"],
-            json!("https://api.stepfun.ai")
-        );
-        assert!(note.contains("https://api.stepfun.com"));
-        assert!(note.contains("https://api.stepfun.ai"));
-        assert!(catalog_failure_hint.contains("https://api.stepfun.com"));
-        assert!(catalog_failure_hint.contains("https://api.stepfun.ai"));
-        assert!(request_failure_hint.contains("https://api.stepfun.com"));
-        assert!(request_failure_hint.contains("https://api.stepfun.ai"));
-    }
-
-    #[test]
-    fn custom_models_endpoint_avoids_double_v1_suffix() {
-        let config = ProviderConfig {
-            kind: ProviderKind::Custom,
-            base_url: "https://example.test/v1".to_owned(),
-            ..ProviderConfig::default()
-        };
-
-        assert_eq!(
-            config.endpoint(),
-            "https://example.test/v1/chat/completions"
-        );
-        assert_eq!(config.models_endpoint(), "https://example.test/v1/models");
-    }
-
-    #[test]
-    fn explicit_api_key_binding_beats_default_oauth_fallback() {
-        let mut env = ScopedEnv::new();
-        env.set("OPENAI_API_KEY", "api-key-wins");
-        env.set("OPENAI_CODEX_OAUTH_TOKEN", "oauth-fallback-should-not-win");
-
-        let config = ProviderConfig {
-            kind: ProviderKind::Openai,
-            api_key: Some(SecretRef::Inline("${OPENAI_API_KEY}".to_owned())),
-            ..ProviderConfig::default()
-        };
-
-        assert_eq!(config.oauth_access_token(), None);
-        assert_eq!(config.api_key().as_deref(), Some("api-key-wins"));
-        assert_eq!(
-            config.resolved_auth_secret().as_deref(),
-            Some("api-key-wins")
-        );
-        assert_eq!(
-            config.authorization_header().as_deref(),
-            Some("Bearer api-key-wins")
-        );
-    }
-
-    #[test]
-    fn explicit_api_key_env_field_beats_default_oauth_fallback() {
-        let mut env = ScopedEnv::new();
-        env.set("OPENAI_API_KEY", "api-key-wins");
-        env.set("OPENAI_CODEX_OAUTH_TOKEN", "oauth-fallback-should-not-win");
-
-        let config: ProviderConfig = toml::from_str(
-            r#"
-kind = "openai"
-api_key_env = "OPENAI_API_KEY"
-"#,
-        )
-        .expect("deserialize provider config");
-
-        assert_eq!(config.oauth_access_token(), None);
-        assert_eq!(config.api_key().as_deref(), Some("api-key-wins"));
-        assert_eq!(
-            config.resolved_auth_secret().as_deref(),
-            Some("api-key-wins")
-        );
-        assert_eq!(
-            config.authorization_header().as_deref(),
-            Some("Bearer api-key-wins")
-        );
-    }
-
-    #[test]
-    fn normalized_for_persistence_canonicalizes_legacy_api_key_env_binding() {
-        let mut config = ProviderConfig::fresh_for_kind(ProviderKind::Openai);
-        config.set_api_key_env(Some("OPENAI_API_KEY".to_owned()));
-
-        let normalized = config.normalized_for_persistence();
-
-        assert_eq!(
-            normalized.api_key,
-            Some(SecretRef::Env {
-                env: "OPENAI_API_KEY".to_owned(),
-            })
-        );
-        assert_eq!(normalized.api_key_env, None);
-        assert_eq!(normalized.oauth_access_token, None);
-        assert_eq!(normalized.oauth_access_token_env, None);
-    }
-
-    #[test]
-    fn normalized_for_persistence_keeps_secret_ref_env_binding_canonical() {
-        let mut config = ProviderConfig::fresh_for_kind(ProviderKind::Openai);
-        config.api_key = Some(SecretRef::Env {
-            env: "OPENAI_API_KEY".to_owned(),
-        });
-
-        let normalized = config.normalized_for_persistence();
-
-        assert_eq!(
-            normalized.api_key,
-            Some(SecretRef::Env {
-                env: "OPENAI_API_KEY".to_owned(),
-            })
-        );
-        assert_eq!(normalized.api_key_env, None);
-    }
-
-    #[test]
-    fn normalized_for_persistence_keeps_implicit_provider_auth_defaults_unset() {
-        let config = ProviderConfig::fresh_for_kind(ProviderKind::Openai);
-
-        let normalized = config.normalized_for_persistence();
-
-        assert_eq!(normalized.api_key, None);
-        assert_eq!(normalized.api_key_env, None);
-        assert_eq!(normalized.oauth_access_token, None);
-        assert_eq!(normalized.oauth_access_token_env, None);
-    }
-
-    #[test]
-    fn canonicalize_configured_auth_env_bindings_rewrites_inline_env_templates() {
-        let mut config = ProviderConfig::fresh_for_kind(ProviderKind::Openai);
-        config.set_api_key_env(Some("OPENAI_API_KEY".to_owned()));
-        config.api_key = Some(SecretRef::Inline("${OPENAI_API_KEY}".to_owned()));
-
-        config.canonicalize_configured_auth_env_bindings();
-
-        assert_eq!(
-            config.api_key,
-            Some(SecretRef::Env {
-                env: "OPENAI_API_KEY".to_owned(),
-            })
-        );
-        assert_eq!(config.api_key_env, None);
-    }
-
-    #[test]
-    fn canonicalize_configured_auth_env_bindings_treats_blank_inline_secret_as_absent() {
-        let mut config = ProviderConfig::fresh_for_kind(ProviderKind::Openai);
-        config.set_api_key_env(Some("OPENAI_API_KEY".to_owned()));
-        config.api_key = Some(SecretRef::Inline("   ".to_owned()));
-
-        config.canonicalize_configured_auth_env_bindings();
-
-        assert_eq!(
-            config.api_key,
-            Some(SecretRef::Env {
-                env: "OPENAI_API_KEY".to_owned(),
-            })
-        );
-        assert_eq!(config.api_key_env, None);
-    }
-
-    #[test]
-    fn fresh_minimax_provider_does_not_seed_hidden_preferred_models() {
-        let config = ProviderConfig::fresh_for_kind(ProviderKind::Minimax);
-
-        assert_eq!(config.model, "auto");
-        assert!(
-            config.preferred_models.is_empty(),
-            "provider defaults should not inject hidden runtime fallback models: {config:#?}"
-        );
-    }
-
-    #[test]
-    fn configured_auto_model_candidates_require_explicit_preferred_models() {
-        let config = ProviderConfig {
-            kind: ProviderKind::Minimax,
-            model: "auto".to_owned(),
-            ..ProviderConfig::default()
-        };
-
-        assert!(
-            config.configured_auto_model_candidates().is_empty(),
-            "auto-model fallback candidates should only exist when the operator configured preferred_models explicitly"
-        );
-    }
-
-    #[test]
-    fn only_reviewed_providers_expose_onboarding_models() {
-        assert_eq!(
-            ProviderKind::Deepseek.recommended_onboarding_model(),
-            Some("deepseek-chat")
-        );
-        assert_eq!(
-            ProviderKind::Minimax.recommended_onboarding_model(),
-            Some("MiniMax-M2.7")
-        );
-        assert_eq!(
-            ProviderKind::Xiaomi.recommended_onboarding_model(),
-            Some("mimo-v2-pro")
-        );
-        assert_eq!(
-            ProviderKind::KimiCoding.recommended_onboarding_model(),
-            None
-        );
-        assert_eq!(ProviderKind::Openai.recommended_onboarding_model(), None);
-    }
-
-    #[test]
-    fn model_catalog_probe_recovery_requires_explicit_model_for_reviewed_auto_provider() {
-        let config = ProviderConfig {
-            kind: ProviderKind::Deepseek,
-            model: "auto".to_owned(),
-            ..ProviderConfig::default()
-        };
-
-        assert_eq!(
-            config.model_catalog_probe_recovery(),
-            ModelCatalogProbeRecovery::RequiresExplicitModel {
-                recommended_onboarding_model: Some("deepseek-chat"),
-            }
-        );
-    }
-
-    #[test]
-    fn model_catalog_probe_recovery_prefers_explicit_runtime_configuration() {
-        let explicit = ProviderConfig {
-            kind: ProviderKind::Deepseek,
-            model: "deepseek-chat".to_owned(),
-            ..ProviderConfig::default()
-        };
-        assert_eq!(
-            explicit.model_catalog_probe_recovery(),
-            ModelCatalogProbeRecovery::ExplicitModel("deepseek-chat".to_owned())
-        );
-
-        let preferred = ProviderConfig {
-            kind: ProviderKind::Deepseek,
-            model: "auto".to_owned(),
-            preferred_models: vec![
-                "deepseek-chat".to_owned(),
-                "deepseek-chat".to_owned(),
-                "deepseek-reasoner".to_owned(),
-            ],
-            ..ProviderConfig::default()
-        };
-        assert_eq!(
-            preferred.model_catalog_probe_recovery(),
-            ModelCatalogProbeRecovery::ConfiguredPreferredModels(vec![
-                "deepseek-chat".to_owned(),
-                "deepseek-reasoner".to_owned(),
-            ])
-        );
-    }
-}
+mod tests;
